move Aura Router path
